+++ src/misc/lv_misc.mk
@@ -11,6 +11,7 @@ CSRCS += lv_log.c
 CSRCS += lv_lru.c
 CSRCS += lv_math.c
 CSRCS += lv_mem.c
+CSRCS += lv_mem_builtin.c
 CSRCS += lv_printf.c
 CSRCS += lv_style.c
 CSRCS += lv_style_gen.c
